fixed types

+++ rentacar/src/main/java/com/omertdemirel/rentacar/business/abstracts/CarService.java
@@ -1,4 +1,5 @@
 package com.omertdemirel.rentacar.business.abstracts;
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.data.domain.Sort;
@@ -22,7 +23,7 @@ public interface CarService {
 
 	Result update(UpdateCarRequest updateCarRequest);
 
-	DataResult<List<ListCarDto>> getByDailyPriceLessThan(double maxDailyPrice);
+	DataResult<List<ListCarDto>> getByDailyPriceLessThan(BigDecimal maxDailyPrice);
 
 	DataResult<List<ListCarDto>> getAllPaged(int pageNo, int pageSize);
